Activate clippy checks on exported APIs
Today I learned about the `avoid-breaking-exported-api = false` clippy
config.

https://doc.rust-lang.org/stable/clippy/lint_configuration.html?highlight=avoid-breaking-exported-api#avoid-breaking-exported-api

While a clippy warning probably is not a good justification for a breaking
change, such warnings can be useful to avoid issues in new APIs.

The number of issues clipply flags for existing APIs is managable, so
I propose adding that config, and add matching allow annotations to
override warnings in the current API.