Add `start_dma` method to `AdcFifoBuilder`
The method first enables the FIFO, sets FCS.THRESH = 1 and FCS.DREQ_EN = 1.

Afterwards the given closure is called with an appropriate
`dma::ReadTarget`, yielding either 8-bit or 16-bit samples (depending
on `FCS.SHIFT`).

The closure must start a DMA transfer and return it.

Finally CS.START_MANY is set, and a `AdcFifo` together with the DMA
transfer is returned.