Avoid missing docs warning on pub extern crate
Due to https://github.com/rust-lang/rust/issues/112308, the public
re-export of the pac will trigger a missing docs warning in rust
1.71.0.

This is bogus, as the doc string would not be shown in the rendered
docs anyway. However, the warning would be annoying, so add a short
doc string to the item.