Use same USB_PLL default setting as the C SDK
The previously used settings used a VCO clock of 480MHz, which is
outside the valid range (750MHz - 1600MHz) given by the datasheet since
revision 1.8.

There are several valid VCO values, and there is a tradeoff between
lower jitter and lower power consumption. So it's not obvious which
value would be the best default.

Instead of doing our own guesswork, use the same value as the C SDK:
https://github.com/raspberrypi/pico-sdk/pull/869/files

This patch only changes the default value, but doesn't update the
input validation. So it's still possible to manually specify a VCO
frequency outside the updated range, as long as it's in the more lenient
range specified by older datasheet revisions, 400MHz - 1600MHz.