Disable clippy lint needless_pass_by_ref_mut
We sometimes use &mut to guarantee exclusive access to some ressource,
where the compiler can't see that such a guarantee is needed for
soundness. In this case, warnings about needles &mut references should
not be generated.

As the lint is not yet known in stable rust, also disable warnings about
unknown lints, so the code compiles without warnings in both current
stable and newer rust versions without warnings.