Add from_installed_program with correct shift direction
The existing function from_program uses the "wrong" shift direction,
the default is different from what the RP2040 does. This is confusing
and can lead to difficult to debug errors. (This is how this difference
was found.)

To avoid a subtle breaking change by just fixing the default, introduce
a new function from_installed_program with the correct default value,
and deprecate the old one.