Replace asm macros by rust macros
According to https://doc.rust-lang.org/stable/reference/inline-assembly.html#directives-support
the assembly directive `.macro` is not part of the supported subset.

The reference further explains: "The result of using other directives is
assembler-specific (and may cause an error, or may be accepted as-is)"

So while it obviously works right now, it might cause issues in future
versions of rust. Not sure if this would be avoided as a breaking
change. To be on the safe side, replace asm macros by rust macros.