Fix float_to_fix64 return value & docs for f32 trig functions
The current return argument for float_to_fix64 doesn't
make sense and is at odds with the definition in the SDK.
Replace it with the correct type.

Similarly, the function descriptions for fcos, fsin, and ftan
give a valid range inconsistent with the values given
on page 140 of thedatasheet. This corrects them.