i2c: Handle the case where a restart is issued right after the address
One case is on read operations for 10bit I2C.
First a write is issued with the address, then a restart and a the msb of
the address is resent with this time the read flag.
Because there is no data phase, the peripheral entered the active state but
never switched to either Read or Write.