Remove code setting IC_DATA_CMD.RESTART bit (#827)
* Remove code setting IC_DATA_CMD.RESTART bit

According to the datasheet, a restart is automatically triggered "if
the transfer direction is changing from the previous command". The
RESTART bit is only needed to trigger an unconditional RESTART sequence.

According to the contract of I2c::transaction, "Data from
adjacent operations of the same type are sent after each other without
an SP or SR" and "Between adjacent operations of a different type an SR
and SAD+R/W is sent". This looks like it is exactly what the hardware
provides out of the box.

(See: https://docs.rs/embedded-hal/1.0.0/ebedded_hal/i2c/trait.I2c.html#tymethod.transaction)

* Increase rtt buffer size for on-target tests

Some of the panic messages are longer than the default buffer, causing a lockup.

* Fix on-target tests

* Also apply changes to rp235x-hal