Remove elf2uf2-rs in favour of picotool
elf2uf2-rs seems to be unmaintained, and as of Rust 1.89 rejects valid
ELF files due to an overly strict check of the ELF header. Picotool is
maintained by Raspberry Pi and does not have this problem.

This commit changes guide-level explanation in the root and RP2040
examples READMEs, and the runner in .cargo/config.toml files, to
picotool.

Some wording around the flashing process as "creating a UF2 file" was
changed as picotool can flash directly from an ELF file with the
appropriate command-line flag. Information about deliberately creating a
UF2 file has been kept.

In addition, the root README now directs people to rp-binary-info rather
than stating that picotool-compatible "binary info" is unsupported.